@@ -255,7 +255,7 @@ namespace ntbs_service.DataAccess
             // We can change the EF behaviour by setting UseRelationalNulls to false, however this can only be done
             // globally and in most situations we probably want the default behaviour.
             // See https://docs.microsoft.com/en-us/dotnet/api/microsoft.entityframeworkcore.infrastructure.relationaldbcontextoptionsbuilder-2.userelationalnulls?view=efcore-5.0
-            return await _context.NotificationAndDuplicateIds.FromSql(@"SELECT N1.[NotificationId] AS [NotificationId], N2.[NotificationId] AS [DuplicateId]
+            return await _context.NotificationAndDuplicateIds.FromSqlRaw(@"SELECT N1.[NotificationId] AS [NotificationId], N2.[NotificationId] AS [DuplicateId]
 FROM [Notification] AS N1
 LEFT JOIN [Patients] AS P1 ON N1.[NotificationId] = P1.[NotificationId]
 CROSS JOIN [Notification] AS N2
