@@ -310,9 +310,8 @@ namespace ntbs_service.DataAccess
 
         public async Task<NotificationGroup> GetNotificationGroupAsync(int notificationId)
         {
-            return await _context.Notification
-                .Where(n => n.NotificationId == notificationId)
-                .Select(n => n.Group)
+            return await _context.NotificationGroup
+                .Where(g => g.Notifications.Select(e => e.NotificationId).Contains(notificationId))
                 .Include(g => g.Notifications)
                     .ThenInclude(n => n.PatientDetails)
                         .ThenInclude(p => p.PostcodeLookup)
@@ -327,6 +326,7 @@ namespace ntbs_service.DataAccess
                 .Include(g => g.Notifications)
                     .ThenInclude(n => n.HospitalDetails)
                         .ThenInclude(e => e.CaseManager)
+                .AsSplitQuery()
                 .SingleOrDefaultAsync();
         }
 
@@ -340,7 +340,8 @@ namespace ntbs_service.DataAccess
                 .Include(n => n.SocialContextAddresses)
                 .Include(n => n.SocialContextVenues)
                 .Include(n => n.TreatmentEvents)
-                    .ThenInclude(t => t.TreatmentOutcome);
+                    .ThenInclude(t => t.TreatmentOutcome)
+                .AsSplitQuery();
         }
 
         // Gets Notification model with basic information for use in notifications homepage lists.
