@@ -72,8 +72,8 @@ namespace ntbs_service.Models.SeedData
                 new VenueType { VenueTypeId = 61, Name = "Mental health rehabilitation centre", Category = "Treatment and rehab" },
                 new VenueType { VenueTypeId = 62, Name = "Other treatment or rehab centre", Category = "Treatment and rehab" },
                 new VenueType { VenueTypeId = 63, Name = "Crisis centre or refuge", Category = "Health and care" },
-                new VenueType { VenueTypeId = 64, Name = "Initial accommodation centre ", Category = "Residential" },
-                new VenueType { VenueTypeId = 65, Name = "Dispersal accommodation ", Category = "Residential" },
+                new VenueType { VenueTypeId = 64, Name = "Initial accommodation centre ", Category = "Residential" },
+                new VenueType { VenueTypeId = 65, Name = "Dispersal accommodation ", Category = "Residential" },
                 new VenueType { VenueTypeId = 66, Name = "Homeless shelter", Category = "Residential" },
                 new VenueType { VenueTypeId = 67, Name = "Squat", Category = "Residential" },
                 new VenueType { VenueTypeId = 68, Name = "Care home", Category = "Residential" },
