@@ -221,7 +221,7 @@ namespace ntbs_service.Services
                     }
                 }
             } 
-            catch(Exception ex) {
+            catch(Exception) {
 
             }
 
